move validation helper functions to separate file
